feat: move WearableBodyShape to BodyShape within item
