+++ src/platform/item/wearable/index.ts
@@ -2,5 +2,5 @@ export { I18N } from '../i18n'
 export { Locale } from '../locale'
 export { Metrics } from '../metrics'
 export { WearableRepresentation } from './representation'
-export { WearableBodyShape } from './wearable-body-shape'
+export { WearableBodyShape } from '../body-shape'
 export { Wearable, ThirdPartyWearable, StandardWearable } from './wearable'
